+++ src/main/java/name/valery1707/megatel/sorm/api/http/HttpDto.java
@@ -25,8 +25,8 @@ public class HttpDto {
 	private String referrer;
 	private String userAgent;
 
-	private Integer requestBodyLen;
-	private Integer responseBodyLen;
+	private Long requestBodyLen;
+	private Long responseBodyLen;
 
 	private Integer statusCode;
 	private String statusMsg;
@@ -140,19 +140,19 @@ public class HttpDto {
 		this.userAgent = userAgent;
 	}
 
-	public Integer getRequestBodyLen() {
+	public Long getRequestBodyLen() {
 		return requestBodyLen;
 	}
 
-	public void setRequestBodyLen(Integer requestBodyLen) {
+	public void setRequestBodyLen(Long requestBodyLen) {
 		this.requestBodyLen = requestBodyLen;
 	}
 
-	public Integer getResponseBodyLen() {
+	public Long getResponseBodyLen() {
 		return responseBodyLen;
 	}
 
-	public void setResponseBodyLen(Integer responseBodyLen) {
+	public void setResponseBodyLen(Long responseBodyLen) {
 		this.responseBodyLen = responseBodyLen;
 	}
